Correções de bugs e add comentários

+++ interacoes_rede.php
@@ -7,42 +7,73 @@
 
   $cod_postagem = $_POST["cod_postagem"];
   $email = $_SESSION["email"];
-  $situacao = $_POST["situacao"];
+  $data = date('Y-m-d');
+  $hora = date('H:i:s');
+  $acao = $_POST["acao"];
 
-  if($situacao == 1){
-    $data = date('Y-m-d');
-    $hora = date('H:i:s');
-    $insert = "INSERT INTO curtida(
+  if($acao == 1){
+    $situacao = $_POST["situacao"];
+
+    if($situacao == 1){
+      $insert = "INSERT INTO curtida(
+        cod_postagem,
+        hora,
+        data,
+        email_usuario
+      ) VALUES(
+          '$cod_postagem',
+          '$hora',
+          '$data',
+          '$email'
+      )";
+      
+      if(mysqli_query($conexao,$insert)){
+      $erro = 0;
+      }        
+      else{
+      $erro = 1;
+      }
+    }
+    else{
+  
+      $delete = "DELETE FROM curtida WHERE cod_postagem = '$cod_postagem' AND email_usuario = '$email' ";
+  
+      if(mysqli_query($conexao,$delete)){
+      $erro = 0;
+      }        
+      else{
+      $erro = 1;
+      }
+      
+    }
+  }
+  else if($acao == 2){  
+    $conteudo = $_POST["conteudo"];
+
+    $insert = "INSERT INTO comentario(
       cod_postagem,
       hora,
       data,
-      email_usuario
+      email_usuario,
+      conteudo
     ) VALUES(
         '$cod_postagem',
         '$hora',
         '$data',
-        '$email'
+        '$email',
+        '$conteudo'
     )";
-    
+
     if(mysqli_query($conexao,$insert)){
-    $erro = 0;
+      $erro = 0;
     }        
     else{
-    $erro = 1;
+      $erro = 1;
     }
-  }
-  else{
 
-    $delete = "DELETE FROM curtida WHERE cod_postagem = '$cod_postagem' AND email_usuario = '$email' ";
-
-    if(mysqli_query($conexao,$delete)){
-    $erro = 0;
-    }        
-    else{
-    $erro = 1;
-    }
-    
   }
+
+
 
   echo $erro;
 ?>
